Move component interfaces to domain
