+++ app/src/main/java/com/github/saikcaskey/pokertracker/domain/components/MainComponent.kt
@@ -1,8 +1,7 @@
-package com.github.saikcaskey.pokertracker.presentation.main
+package com.github.saikcaskey.pokertracker.domain.components
 
 import com.arkivanov.decompose.router.pages.ChildPages
 import com.arkivanov.decompose.value.Value
-import com.github.saikcaskey.pokertracker.shared.presentation.main.MainPagerPageComponent
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.serialization.Serializable
