+++ app/src/main/java/com/github/saikcaskey/pokertracker/domain/components/MainPagerPageComponent.kt
@@ -1,0 +1,3 @@
+package com.github.saikcaskey.pokertracker.domain.components
+
+interface MainPagerPageComponent
