+++ app/src/main/java/com/github/saikcaskey/pokertracker/domain/components/RootComponent.kt
@@ -1,18 +1,7 @@
-package com.github.saikcaskey.pokertracker.shared.presentation.root
+package com.github.saikcaskey.pokertracker.domain.components
 
 import com.arkivanov.decompose.router.stack.ChildStack
 import com.arkivanov.decompose.value.Value
-import com.github.saikcaskey.pokertracker.presentation.event.EventDetailComponent
-import com.github.saikcaskey.pokertracker.presentation.event.InsertEventComponent
-import com.github.saikcaskey.pokertracker.presentation.event.ViewEventsComponent
-import com.github.saikcaskey.pokertracker.presentation.expense.ExpenseDetailComponent
-import com.github.saikcaskey.pokertracker.presentation.expense.InsertExpenseComponent
-import com.github.saikcaskey.pokertracker.shared.presentation.expense.ViewExpensesComponent
-import com.github.saikcaskey.pokertracker.presentation.main.MainComponent
-import com.github.saikcaskey.pokertracker.presentation.planner.DayDetailComponent
-import com.github.saikcaskey.pokertracker.shared.presentation.venue.InsertVenueComponent
-import com.github.saikcaskey.pokertracker.shared.presentation.venue.VenueDetailComponent
-import com.github.saikcaskey.pokertracker.presentation.venue.ViewVenuesComponent
 
 interface RootComponent {
