+++ app/src/main/java/com/github/saikcaskey/pokertracker/domain/components/ViewExpensesComponent.kt
@@ -1,4 +1,4 @@
-package com.github.saikcaskey.pokertracker.shared.presentation.expense
+package com.github.saikcaskey.pokertracker.domain.components
 
 import com.github.saikcaskey.pokertracker.domain.models.Expense
 import kotlinx.coroutines.flow.StateFlow
