--- app/src/main/java/com/github/saikcaskey/pokertracker/presentation/main/MainPagerPageComponent.kt
@@ -1,3 +1,0 @@
-package com.github.saikcaskey.pokertracker.shared.presentation.main
-
-interface MainPagerPageComponent
